requirements.txt
fastapi
uvicorn
pywebview
pydantic
requests
typer